[AspNetCore] Simplify benchmarks
Have a single benchmark that uses the test host and remove various cruft.
