+++ Directory.Packages.props
@@ -119,6 +119,7 @@
     <PackageVersion Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="8.0.25" />
     <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.25" />
     <PackageVersion Include="Microsoft.AspNetCore.SignalR.Client" Version="8.0.25" />
+    <PackageVersion Include="Microsoft.AspNetCore.TestHost" Version="8.0.25" />
     <PackageVersion Include="Microsoft.Data.SqlClient" Version="7.0.0" />
     <PackageVersion Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.25" />
     <PackageVersion Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.25" />
@@ -151,11 +152,13 @@
     <PackageVersion Update="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="9.0.14" />
     <PackageVersion Update="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.14" />
     <PackageVersion Update="Microsoft.AspNetCore.SignalR.Client" Version="9.0.14" />
+    <PackageVersion Update="Microsoft.AspNetCore.TestHost" Version="9.0.14" />
   </ItemGroup>
   <ItemGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net10.0'))" Label="Non-production packages, such as for examples and tests, for .NET 10">
     <PackageVersion Update="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="10.0.5" />
     <PackageVersion Update="Microsoft.AspNetCore.Mvc.Testing" Version="10.0.5" />
     <PackageVersion Update="Microsoft.AspNetCore.SignalR.Client" Version="10.0.5" />
+    <PackageVersion Update="Microsoft.AspNetCore.TestHost" Version="10.0.5" />
   </ItemGroup>
 
 </Project>
