+++ test/OpenTelemetry.Instrumentation.AspNetCore.Benchmarks/Program.cs
@@ -2,10 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using BenchmarkDotNet.Running;
+using OpenTelemetry.Instrumentation.AspNetCore.Benchmarks;
 
-namespace OpenTelemetry.Instrumentation.AspNetCore.Benchmarks;
-
-internal class Program
-{
-    private static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
-}
+BenchmarkSwitcher.FromAssembly(typeof(AspNetCoreBenchmarks).Assembly).Run(args);
